Redid sign in/up UI. Added better form validation
