+++ CoffeeLog/Screens/Account/AccountView.swift
@@ -1,0 +1,20 @@
+////
+////  AccountView.swift
+////  CoffeeLog
+////
+////  Created by Tim West on 9/11/23.
+////
+//
+//import SwiftUI
+//
+//struct AccountView: View {
+//    var body: some View {
+//
+//    }
+//}
+//
+//struct AccountView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        AccountView()
+//    }
+//}
